packages/core/pyproject.toml
[tool.poetry]
name = "zenus-core"
version = "0.6.0"
description = "Zenus Core - Intent Execution Engine"
authors = ["Zenus Team"]
readme = "README.md"
license = "MIT"
packages = [{include = "zenus_core", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.0"
watchdog = "^3.0.0"
requests = "^2.31.0"
openai = "^1.0.0"
anthropic = "^0.47.0"
rich = "^13.7.0"
psutil = "^5.9.0"
playwright = {version = "^1.40.0", optional = true}
prompt-toolkit = "^3.0.52"
pyautogui = "^0.9.54"
pillow = "^12.1.1"
matplotlib = "^3.8.0"
numpy = "^1.24.0"
zenus-visualization = {path = "../visualization", develop = true}

[tool.poetry.extras]
browser = ["playwright"]
all = ["playwright"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
hypothesis = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
